Create a separate docker image for dream3d runtime
This image is much smaller and better to transport, but it
still requires qtgui, which (including its dependencies) is
around 300 megabytes. We could reduce the size substantially
if we could figure out how to remove qtgui.

Anyways, this image seems to work well other than the problem
that the file output is owned by the root.

Here is an example of how to run it:

docker run -v $PWD:/data -w /data psavery/dream3d input.json